Crate libra_network_address

Structs

DnsName
A minimally parsed DNS name. We don't really do any checking other than enforcing:

EmptyError

NetworkAddress
Libra NetworkAddress is a compact, efficient, self-describing and future-proof network address represented as a stack of protocols. Essentially libp2p's multiaddr but using lcs to describe the binary format.

RawNetworkAddress
A RawNetworkAddress is the serialized, unverified, on-chain representation of a NetworkAddress. Specifically, a RawNetworkAddress is usually an lcs-serialized NetworkAddress.

Enums

ParseError
Possible errors when parsing a human-readable NetworkAddress.

Protocol
A single protocol in the NetworkAddress protocol stack.